refactor: rename paths
