--- packages/game/src/command.ts
@@ -1,6 +1,0 @@
-export interface Command {
-  command: string
-  output: string
-  error: string
-  pwd: string
-}
