+++ packages/game/src/condition/CommandMatcher.ts
@@ -1,5 +1,5 @@
-import { Command } from '../command.js'
-import { Matcher } from './matcher.js'
+import { Command } from '../types.js'
+import { Matcher } from './Matcher.js'
 
 export class CommandMatcher implements Matcher {
   private matcher: RegExp
@@ -7,7 +7,7 @@ export class CommandMatcher implements Matcher {
     this.matcher = new RegExp(command, flags)
   }
 
-  match(command: Partial<Command>): boolean {
+  match(command: Command): boolean {
     return !!command.command && this.matcher.test(command.command)
   }
 }
