+++ packages/game/src/condition/Condition.ts
@@ -1,21 +1,12 @@
 /* eslint-disable max-classes-per-file */
-import { Command } from '../command.js'
-import { PwdMatcher } from './pwdMatcher.js'
-import { CommandMatcher } from './commandMatcher.js'
-import { ErrorMatcher } from './errorMatcher.js'
-import { Matcher } from './matcher.js'
-import { OutputMatcher } from './outputMatcher.js'
-import { OrMatcher } from './orMatcher.js'
-import { NotMatcher } from './notMatcher.js'
-
-export interface ICondition {
-  command?: string
-  output?: string
-  error?: string
-  pwd?: string
-  or?: ICondition[]
-  not?: ICondition
-}
+import { Command, ICondition } from '../types.js'
+import { PwdMatcher } from './PwdMatcher.js'
+import { CommandMatcher } from './CommandMatcher.js'
+import { ErrorMatcher } from './ErrorMatcher.js'
+import { Matcher } from './Matcher.js'
+import { OutputMatcher } from './OutputMatcher.js'
+import { OrMatcher } from './OrMatcher.js'
+import { NotMatcher } from './NotMatcher.js'
 
 export class Condition {
   private matchers: Matcher[] = []
@@ -32,7 +23,7 @@ export class Condition {
     this.matchers.push(new ErrorMatcher(error)) // always check unexpected error
   }
 
-  match(command: Partial<Command>): boolean {
+  match(command: Command): boolean {
     return this.matchers.every((matcher) => {
       return matcher.match(command)
     })
