+++ packages/game/src/condition/Matcher.ts
@@ -1,0 +1,5 @@
+import { Command } from '../types.js'
+
+export interface Matcher {
+  match(command: Command): boolean
+}
