+++ packages/game/src/condition/NotMatcher.ts
@@ -1,10 +1,10 @@
-import { Command } from '../command.js'
-import { Matcher } from './matcher.js'
+import { Command } from '../types.js'
+import { Matcher } from './Matcher.js'
 
 export class NotMatcher implements Matcher {
   constructor(private condition: Matcher) {}
 
-  match(command: Partial<Command>): boolean {
+  match(command: Command): boolean {
     return !this.condition.match(command)
   }
 }
