+++ packages/game/src/condition/OrMatcher.ts
@@ -1,10 +1,10 @@
-import { Command } from '../command.js'
-import { Matcher } from './matcher.js'
+import { Command } from '../types.js'
+import { Matcher } from './Matcher.js'
 
 export class OrMatcher implements Matcher {
   constructor(private conditions: Matcher[]) {}
 
-  match(command: Partial<Command>): boolean {
+  match(command: Command): boolean {
     return this.conditions.some((condition) => condition.match(command))
   }
 }
