+++ packages/game/src/condition/OutputMatcher.ts
@@ -1,5 +1,5 @@
-import { Command } from '../command.js'
-import { Matcher } from './matcher.js'
+import { Command } from '../types.js'
+import { Matcher } from './Matcher.js'
 
 export class OutputMatcher implements Matcher {
   private matcher: RegExp
@@ -7,7 +7,7 @@ export class OutputMatcher implements Matcher {
     this.matcher = new RegExp(pattern, args)
   }
 
-  match(command: Partial<Command>): boolean {
+  match(command: Command): boolean {
     return !!command.output && this.matcher.test(command.output.trim())
   }
 }
