+++ packages/game/src/condition/PwdMatcher.ts
@@ -1,10 +1,10 @@
-import { Command } from '../command.js'
-import { Matcher } from './matcher.js'
+import { Command } from '../types.js'
+import { Matcher } from './Matcher.js'
 
 export class PwdMatcher implements Matcher {
   constructor(private pwd: string) {}
 
-  match(command: Partial<Command>): boolean {
+  match(command: Command): boolean {
     return command.pwd === this.pwd
   }
 }
