--- packages/game/src/condition/matcher.ts
@@ -1,5 +1,0 @@
-import { Command } from '../command.js'
-
-export interface Matcher {
-  match(command: Partial<Command>): boolean
-}
