+++ packages/game/src/types.ts
@@ -1,0 +1,15 @@
+export interface Command {
+  command?: string
+  output?: string
+  error?: string
+  pwd?: string
+}
+
+export interface ICondition {
+  command?: string
+  output?: string
+  error?: string
+  pwd?: string
+  or?: ICondition[]
+  not?: ICondition
+}
